requirements.txt
django
pillow
psycopg2
django-friendship
djangorestframework